- fixed return type of filter_move_all() - remove finished chunks from the filter-queue when we moved all the   chunks to the chunk-queue
git-svn-id: svn://svn.lighttpd.net/lighttpd/trunk@1888 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/filter.h
@@ -37,6 +37,6 @@ LI_API filter * filter_chain_create_filter(filter_chain *chain, int id);
 LI_API filter * filter_chain_get_filter(filter_chain *chain, int id);
 LI_API void filter_chain_remove_filter(filter_chain *chain, filter *fl);
 
-LI_API int filter_chain_copy_output(filter_chain *chain, chunkqueue *out);
+LI_API off_t filter_chain_copy_output(filter_chain *chain, chunkqueue *out);
 
 #endif
